+++ backend/MpumalangaAssetManagement/MAM.DataAccess/Tables/HiredProperty.cs
@@ -13,14 +13,14 @@ namespace MAM.DataAccess.Tables
         public string BuildingCondition { get; set; }
         public DateTime StartingDate { get; set; }
         public DateTime TerminationDate { get; set; }
-        public decimal MonthlyRental { get; set; }
+        public double MonthlyRental { get; set; }
         public string Town { get; set; }
         public string Status { get; set; }
         public string UserDepartment { get; set; }
         public string LandlandAgentName { get; set; }
         public int NumberofStuff { get; set; }
-        public int EscalationRate { get; set; }
-        public decimal Area { get; set; }
+        public double EscalationRate { get; set; }
+        public double Area { get; set; }
         public string Address { get; set; }
         public bool IsDeteted { get; set; }
         public int CreatedUserId { get; set; }
